+++ client/src/components/ui/provider.tsx
@@ -1,12 +1,16 @@
 'use client';
 
-import { ChakraProvider, defaultSystem } from '@chakra-ui/react';
+import { ChakraProvider } from '@chakra-ui/react';
+import { ThemePresetProvider } from '@/features/theme/theme-provider';
+import { system } from '@/theme/system';
 import { ColorModeProvider, type ColorModeProviderProps } from './color-mode';
 
-export function Provider(props: ColorModeProviderProps) {
+export function Provider({ children, ...props }: ColorModeProviderProps) {
   return (
-    <ChakraProvider value={defaultSystem}>
-      <ColorModeProvider {...props} />
+    <ChakraProvider value={system}>
+      <ColorModeProvider {...props}>
+        <ThemePresetProvider>{children}</ThemePresetProvider>
+      </ColorModeProvider>
     </ChakraProvider>
   );
 }